# api/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Qeet Pay API
  description: >-
    Qeet Pay — India-first payments, billing and GST infrastructure (UPI / Cards / NACH). Programmatic `/v1` calls
    authenticate with a `qp_live_` / `qp_test_` API key via the `X-Api-Key` header; money is always in integer minor
    units (paise).
  contact:
    name: Qeet Group
    url: https://qeet.in
  license:
    name: Proprietary
    url: https://qeet.in
  version: v1
servers:
  - url: https://api.pay.qeet.in
    description: Production
  - url: https://api.pay.staging.qeet.in
    description: Staging
  - url: http://localhost:4201
    description: Local (docker compose)
security:
  - ApiKeyAuth: []
...
paths:
  /v1/payments/providers/{provider}/cost:
    put:
      tags:
        - Orchestration
      operationId: setCost
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CostRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ScorecardView'
...
  /v1/payments:
    get:
      tags:
        - Payments
      operationId: list_6
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PaymentView'
    post:
      tags:
        - Payments
      operationId: create_3
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaymentView'
  /v1/payments/{id}/refund:
    post:
      tags:
        - Payments
      operationId: refund
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /v1/payments/{id}/capture:
    post:
      tags:
        - Payments
      operationId: capture
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: object
  /v1/payments/razorpay/webhook:
    post:
      tags:
        - Payments
      operationId: handleWebhook
      parameters:
        - name: X-Razorpay-Signature
          in: header
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: string
        required: true
      responses:
        '200':
          description: OK
...
  /v1/payments/{id}:
    get:
      tags:
        - Payments
      operationId: get_7
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaymentView'
  /v1/payments/{id}/refunds:
    get:
      tags:
        - Payments
      operationId: refunds
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RefundView'
...
  /v1/dunning/subscriptions/{subscriptionId}/attempts:
    get:
      tags:
        - Dunning
      operationId: attempts
      parameters:
        - name: subscriptionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AttemptView'
...
components:
  schemas:
    CostRequest:
      type: object
      properties:
        costBps:
          type: integer
          format: int32
          maximum: 10000
          minimum: 0
    ScorecardView:
      type: object
      properties:
        provider:
          type: string
        attempts:
          type: integer
          format: int64
        successes:
          type: integer
          format: int64
        failures:
          type: integer
          format: int64
        authRate:
          type: number
          format: double
        consecutiveFailures:
          type: integer
          format: int32
        costBps:
          type: integer
          format: int32
        health:
          type: string
        lastOutcomeAt:
          type: string
          format: date-time
...
    CreatePaymentRequest:
      type: object
      properties:
        amountMinor:
          type: integer
          format: int64
        currency:
          type: string
          minLength: 1
        method:
          type: string
          enum:
            - UPI
            - CARD
            - NET_BANKING
            - WALLET
        description:
          type: string
        simulateFailure:
          type: boolean
      required:
        - currency
        - method
    PaymentView:
      type: object
      properties:
        id:
          type: string
        amountMinor:
          type: integer
          format: int64
        currency:
          type: string
        method:
          type: string
        status:
          type: string
        providerPaymentId:
          type: string
        ledgerEntryId:
          type: string
    RefundRequest:
      type: object
      properties:
        amountMinor:
          type: integer
          format: int64
        reason:
          type: string
...
    AttemptView:
      type: object
      properties:
        id:
          type: string
        subscriptionId:
          type: string
        attemptNumber:
          type: integer
          format: int32
        scheduledAt:
          type: string
          format: date-time
        attemptedAt:
          type: string
          format: date-time
        result:
          type: string
        failureReason:
          type: string
        failureCategory:
          type: string
        recommendedDelayHours:
          type: integer
          format: int32
        recommendedChannels:
          type: string
        classificationRationale:
          type: string
...
    RefundView:
      type: object
      properties:
        id:
          type: string
        paymentId:
          type: string
        amountMinor:
          type: integer
          format: int64
        status:
          type: string
        ledgerEntryId:
          type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: Programmatic API key (`qp_live_…` / `qp_test_…`) presented as the `X-Api-Key` header.
      name: X-Api-Key
      in: header
...